server/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
faster-whisper==1.1.0
requests==2.32.3
google-generativeai==0.8.6
python-dotenv==1.1.0
